Revert "cmdline, fastboot: Use ACS_VERSION instead of U_BOOT_DATE_TIME"
This reverts commit 2de5336ab9d0a4a41e21babedb0f843e5d21c533.

For some reason (not completely understood currently), the change in
bootloader version name makes the system prevent switching slots after
OTA. The OTA seems to succeed (Updater app reports OK and post-boot
toast message state "upgrade successful"), but the slot is never changed.
AVB2 or VAB in the bootloader might be the culprit...

This commit is reverted because the OTA process also updates the
booloader, so no bootloader/Linux/Android incompatibilities can be introduced
by the OTA process. (Previously we belived that OTA did not update the
bootloader and that this couldd be a source for bootloader/Linux/Android
incompatibilities)